+++ piecrust-uplink/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2023-07-19
+
 ### Added
 
 - Add more comprehensive documentation of the whole crate [#189]&[#190]
@@ -111,7 +113,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#136]: https://github.com/dusk-network/piecrust/issues/136
 
 <!-- VERSIONS -->
-[Unreleased]: https://github.com/dusk-network/piecrust/compare/uplink-0.6.1...HEAD
+[Unreleased]: https://github.com/dusk-network/piecrust/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/dusk-network/piecrust/compare/uplink-0.6.1...v0.7.0
 [0.6.1]: https://github.com/dusk-network/piecrust/compare/v0.6.0...uplink-0.6.1
 [0.6.0]: https://github.com/dusk-network/piecrust/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/dusk-network/piecrust/compare/v0.4.0...v0.5.0